Add simple script to get coordinates from RMF files.
This can be used to extract the coordinates of all residues
and beads from an RMF file. Include the output of running it
on the top-scoring model from each of the two clusters generated
in the Nup84 study.

Output is currently in 'human readable' form, but it should be
straightforward to change that to mmCIF or another format.