Add multistate information to the exosome example.
The original exosome modeling publication in fact modeled
two states of the exosome complex, one nuclear-localized
and the other cytoplasm-localized, but due to limitations
of the mmCIF support in IMP only a single state was previously
deposited as mmCIF. IMP's mmCIF support now understands
multiple states, so both of the states are now included
in the exosome example. Closes #9.