Match latest dictionary, remove duplicate crosslinks.
Regenerate all IMP examples using latest IMP, which should
make them compliant with the latest version of the extension
dictionary. This also updates the Mediator example to use
the latest DOI (the latest release of the DOI includes the raw
trajectories from which the final clusters were generated) and
to remove duplicate crosslinks.