Update exosome example with latest IMP
One of the starting models (Rrp46_gfp-m2) for the exosome
uses a different numbering scheme to that of the resulting
IHM models, so we set
ihm_starting_model_details.starting_model_sequence_offset
accordingly. However, IMP was previously ignoring this
offset when reading in the starting model PDB and so using
the wrong sequence range (in fact, an empty one) when
populating ihm_starting_model_coord. This should now be
resolved. Closes #49.